Revert the "ApiModule" move to fix TS errors
