+++ src/query/react/module.ts
@@ -12,6 +12,8 @@ import { capitalize } from '../utils'
 import { safeAssign } from '../tsHelpers'
 import { BaseQueryFn } from '../baseQueryTypes'
 
+import { HooksWithUniqueNames } from './versionedTypes/index'
+
 import {
   useDispatch as rrUseDispatch,
   useSelector as rrUseSelector,
@@ -24,35 +26,45 @@ import { PrefetchOptions } from '../core/module'
 export const reactHooksModuleName = Symbol()
 export type ReactHooksModule = typeof reactHooksModuleName
 
-export interface ReactHooksBaseEndpoints<
-  Definitions extends EndpointDefinitions
-> {
-  /**
-   *  Endpoints based on the input endpoints provided to `createApi`, containing `select`, `hooks` and `action matchers`.
-   */
-  endpoints: {
-    [K in keyof Definitions]: Definitions[K] extends QueryDefinition<
-      any,
-      any,
-      any,
-      any,
-      any
-    >
-      ? QueryHooks<Definitions[K]>
-      : Definitions[K] extends MutationDefinition<any, any, any, any, any>
-      ? MutationHooks<Definitions[K]>
-      : never
+declare module '../apiTypes' {
+  export interface ApiModules<
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    BaseQuery extends BaseQueryFn,
+    Definitions extends EndpointDefinitions,
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    ReducerPath extends string,
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    TagTypes extends string
+  > {
+    [reactHooksModuleName]: {
+      /**
+       *  Endpoints based on the input endpoints provided to `createApi`, containing `select`, `hooks` and `action matchers`.
+       */
+      endpoints: {
+        [K in keyof Definitions]: Definitions[K] extends QueryDefinition<
+          any,
+          any,
+          any,
+          any,
+          any
+        >
+          ? QueryHooks<Definitions[K]>
+          : Definitions[K] extends MutationDefinition<any, any, any, any, any>
+          ? MutationHooks<Definitions[K]>
+          : never
+      }
+      /**
+       * A hook that accepts a string endpoint name, and provides a callback that when called, pre-fetches the data for that endpoint.
+       */
+      usePrefetch<EndpointName extends QueryKeys<Definitions>>(
+        endpointName: EndpointName,
+        options?: PrefetchOptions
+      ): (
+        arg: QueryArgFrom<Definitions[EndpointName]>,
+        options?: PrefetchOptions
+      ) => void
+    } & HooksWithUniqueNames<Definitions>
   }
-  /**
-   * A hook that accepts a string endpoint name, and provides a callback that when called, pre-fetches the data for that endpoint.
-   */
-  usePrefetch<EndpointName extends QueryKeys<Definitions>>(
-    endpointName: EndpointName,
-    options?: PrefetchOptions
-  ): (
-    arg: QueryArgFrom<Definitions[EndpointName]>,
-    options?: PrefetchOptions
-  ) => void
 }
 
 // type RR = typeof import('react-redux')
